Centralize manual schema updates
