+++ schemas/json/base_metadata_schema.json
