+++ schemas/json/path-expansion.schema.json
@@ -1,0 +1,18 @@
+{
+    "properties": {
+        "base_directory": {
+            "type": "string",
+            "format": "directory",
+            "description": "Enter the base directory of your data."
+        },
+        "file_path": {
+            "type": "string",
+            "description": "Enter a format string to locate data.",
+            "placeholder": "{subject_id}_{session_id}_{task_name}/{subject_id}_{session_id}_{task_name}_imec0/{subject_id}_{session_id}_{task_name}_t0.imec0.ap.bin"
+        }
+    },
+    "required": [
+        "base_directory",
+        "file_path"
+    ]
+}
