+++ schemas/project-metadata.schema.ts
@@ -1,0 +1,3 @@
+import projectMetadataSchema from './json/project-metadata.schema.json' assert { type: "json" }
+
+export default projectMetadataSchema
