+++ src/renderer/src/stories/Table.stories.js
@@ -1,6 +1,6 @@
 import { Table } from './Table.js';
 
-import nwbBaseSchema from '../../../../schemas/base_metadata_schema.json'
+import subjectSchema from '../../../../schemas/subject.schema'
 
 export default {
   title: 'Components/Table',
@@ -19,22 +19,10 @@ const data = ['001', '002', '003', '004', '005', '006', '007', '008', '009', '01
   return acc
 }, {})
 
-const schema = {
-  ...nwbBaseSchema.properties.Subject,
-  properties: {
-    sessions: {
-      type: 'array',
-      items: {
-        type: 'string',
-      }
-    },
-    ...nwbBaseSchema.properties.Subject.properties,
-  }
-}
 
 export const Default = Template.bind({});
 Default.args = {
-  schema,
+  schema: subjectSchema,
   data,
   keyColumn: 'subject_id',
   validateOnChange: () => true // Always validate as true
