+++ src/renderer/src/stories/forms/utils.ts
@@ -1,0 +1,6 @@
+export const capitalize = (str: string) => str[0].toUpperCase() + str.slice(1)
+
+
+export const header = (headerStr: string) => {
+    return headerStr.split('_').filter(str => !!str).map(capitalize).join(' ')
+}
